+++ packages/ente-api/src/db/migrations/index.ts
@@ -9,6 +9,7 @@ import { UseEnumType1551372175000 } from "./1551372175000-UseEnumType";
 import { MakeKeyValueStoreTextType1551372598000 } from "./1551372598000-MakeKeyValueStoreTextType";
 import { MakeKeyValueStoreNotNullable1551377199000 } from "./1551377199000-MakeKeyValueStoreNotNullable";
 import { RemoveInconsistentUserAttributes1551534415000 } from "./1551534415000-RemoveInconsistentUserAttributes";
+import { AddNewEntryReasonCategories1551552955000 } from "./1551552955000-AddNewEntryReasonCategories";
 
 export const migrations = [
   MakeSlotDateNullable1544015021618,
@@ -21,5 +22,7 @@ export const migrations = [
   UseEnumType1551372175000,
   MakeKeyValueStoreTextType1551372598000,
   MakeKeyValueStoreNotNullable1551377199000,
+  RemoveInconsistentUserAttributes1551534415000,
+  AddNewEntryReasonCategories1551552955000,
   RemoveInconsistentUserAttributes1551534415000
 ];